+++ pkg/local_object_storage/metabase/lock_test.go
@@ -162,7 +162,7 @@ func TestDB_Lock_Removed(t *testing.T) {
 		}},
 		{name: "with target and GC mark", preset: func(t *testing.T, db *meta.DB) {
 			require.NoError(t, db.Put(&obj))
-			_, _, err := db.MarkGarbage(objAddr)
+			_, err := db.MarkGarbage(objAddr)
 			require.NoError(t, err)
 		}},
 	} {