+++ pkg/local_object_storage/shard/inhume.go
@@ -27,16 +27,21 @@ func (s *Shard) MarkGarbage(addrs ...oid.Address) error {
 		return ErrDegradedMode
 	}
 
-	var err error
-
-	_, _, err = s.metaBase.MarkGarbage(addrs...)
-
+	inhumed, err := s.metaBase.MarkGarbage(addrs...)
 	if err != nil {
 		s.log.Debug("could not mark object to delete in metabase",
 			zap.Error(err),
 		)
 
 		return fmt.Errorf("metabase inhume: %w", err)
+	}
+
+	for _, cnrDiff := range inhumed {
+		cnrStr := cnrDiff.CID.EncodeToString()
+
+		s.addObjectCounter(gcObjType, cnrDiff.NewGarbage)
+		s.addToContainerSize(cnrStr, cnrDiff.PayloadDiff)
+		s.addToPayloadCounter(cnrDiff.PayloadDiff)
 	}
 
 	if s.hasWriteCache() {