+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog for NeoFS Node
 - Support for creation of containers with initial placement policy (#3830)
 - `neofs-lancet meta remove` command (#3891)
 - `neofs-lancet fstree remove` command (#3892)
+- SN has new `root`, `ts`, `lock`, `link`, `gc` object counter metrics (#3740)
 
 ### Fixed
 - Resending the header after chunks have already been sent in object service `Get` handler (#3833)
@@ -40,6 +41,7 @@ Changelog for NeoFS Node
 - 2.18 is the minimum supported version for new objects (#3869)
 - Rename `neofs-lens` into `neofs-lancet` (#3894)
 - SN no longer accepts objects with invocation or verification script bigger than 1KiB (#3887)
+- Every object with non-zero payload is now paid, not only regular ones (#3856)
 
 ### Removed
 - `node.persistent_sessions.path` config option from SN config (#3846)
@@ -47,6 +49,7 @@ Changelog for NeoFS Node
 - `storage.shards.resync_metabase` config option from SN config (#3849)
 - One-time netmap contract placement migration routine for v0.49.0+ releases (#3821)
 - Metabase v5 to v6 and v6 to v7 migrations (#3864)
+- SN's `logic` object counter metric (#3740)
 
 ### Updated
 - `github.com/nspcc-dev/neofs-sdk-go` module to `v1.0.0-rc.17.0.20260320132435-55419a28ca95` (#3785, #3817, #3808, #3830, #3883, #3896, #3887)